fix wonky base_url
